# docs/openapi.yaml
components:
   schemas:
...
      HTTPValidationError:
         properties:
            detail:
               items:
                  $ref: '#/components/schemas/ValidationError'
               title: Detail
               type: array
         title: HTTPValidationError
         type: object
...
      ValidationError:
         properties:
            loc:
               items:
                  anyOf:
                  -  type: string
                  -  type: integer
               title: Location
               type: array
            msg:
               title: Message
               type: string
            type:
               title: Error Type
               type: string
         required:
         - loc
         - msg
         - type
         title: ValidationError
         type: object
info:
   description: "\u0420\u0435\u0444\u0430\u043A\u0442\u043E\u0440\u0438\u0440\u043E\
      \u0432\u0430\u043D\u043D\u044B\u0439 API \u0434\u043B\u044F \u0441\u0438\u0441\
      \u0442\u0435\u043C\u044B \u0443\u043F\u0440\u0430\u0432\u043B\u0435\u043D\u0438\
      \u044F \u043A\u043E\u043B\u043B\u0430\u0436\u0430\u043C\u0438"
   title: CollagrCms API
   version: 0.1.0
openapi: 3.1.0
paths:
...
   /collage/:
      get:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u043A\u043E\u043B\
            \u043B\u0430\u0436 \u0438\u0437 \u0432\u044B\u0431\u0440\u0430\u043D\u043D\
            \u044B\u0445 \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\
            \u0439\n\nArgs:\n    list_id: \u0421\u043F\u0438\u0441\u043E\u043A ID\
            \ \u0438\u0437\u043E\u0431\u0440\u0430\u0436\u0435\u043D\u0438\u0439 (\u043E\
            \u0442 1 \u0434\u043E 12)\n\nReturns:\n    \u0424\u0430\u0439\u043B \u043A\
            \u043E\u043B\u043B\u0430\u0436\u0430 \u0432 \u0444\u043E\u0440\u043C\u0430\
            \u0442\u0435 JPEG"
         operationId: create_collage_collage__get
         parameters:
         -  in: query
            name: list_id
            required: true
            schema:
               items:
                  type: integer
               maxItems: 12
               minItems: 1
               title: List Id
               type: array
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create Collage
         tags:
         - collages
   /collage/batch/:
      get:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u043A\u043E\u043B\
            \u043B\u0430\u0436 \u0438\u0437 \u0442\u0435\u043A\u0443\u0449\u0435\u0433\
            \u043E \u043F\u0430\u043A\u0435\u0442\u0430 \u0442\u043E\u0432\u0430\u0440\
            \u043E\u0432 (\u043F\u043E \u0443\u043C\u043E\u043B\u0447\u0430\u043D\u0438\
            \u044E 12 \u0448\u0442\u0443\u043A)\n\u0418\u0441\u043F\u043E\u043B\u044C\
            \u0437\u0443\u0435\u0442\u0441\u044F \u0434\u043B\u044F \u043A\u043D\u043E\
            \u043F\u043A\u0438 \"\u0414\u0430\u043B\u044C\u0448\u0435\" - \u0441\u043E\
            \u0437\u0434\u0430\u0435\u0442 \u0441\u043B\u0435\u0434\u0443\u044E\u0449\
            \u0438\u0439 \u043A\u043E\u043B\u043B\u0430\u0436 \u0438 \u0441\u0440\u0430\
            \u0437\u0443 \u043E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442\
            \ \u0435\u0433\u043E \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\
            \u0435\u043B\u044E\n\nArgs:\n    batch_size: \u0420\u0430\u0437\u043C\u0435\
            \u0440 \u043F\u0430\u043A\u0435\u0442\u0430 (\u043F\u043E \u0443\u043C\
            \u043E\u043B\u0447\u0430\u043D\u0438\u044E 12)\n    start_index: \u041D\
            \u0430\u0447\u0430\u043B\u044C\u043D\u044B\u0439 \u0438\u043D\u0434\u0435\
            \u043A\u0441 \u0434\u043B\u044F \u043E\u0431\u0440\u0430\u0431\u043E\u0442\
            \u043A\u0438\n\nReturns:\n    \u0424\u0430\u0439\u043B \u043A\u043E\u043B\
            \u043B\u0430\u0436\u0430 \u0434\u043B\u044F \u0441\u043A\u0430\u0447\u0438\
            \u0432\u0430\u043D\u0438\u044F"
         operationId: create_batch_collage_collage_batch__get
         parameters:
         -  in: query
            name: batch_size
            required: false
            schema:
               default: 12
               maximum: 12
               minimum: 1
               title: Batch Size
               type: integer
         -  in: query
            name: start_index
            required: false
            schema:
               default: 0
               minimum: 0
               title: Start Index
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create Batch Collage
         tags:
         - collages
   /collage/batch/all/:
      get:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u0432\u0441\u0435\
            \ \u043A\u043E\u043B\u043B\u0430\u0436\u0438 \u0438\u0437 \u0432\u0441\
            \u0435\u0445 \u0442\u043E\u0432\u0430\u0440\u043E\u0432, \u043E\u0431\u0440\
            \u0430\u0431\u0430\u0442\u044B\u0432\u0430\u044F \u0438\u0445 \u043F\u0430\
            \u043A\u0435\u0442\u0430\u043C\u0438\n\u0418\u0441\u043F\u043E\u043B\u044C\
            \u0437\u0443\u0435\u0442\u0441\u044F \u0434\u043B\u044F \u043A\u043D\u043E\
            \u043F\u043A\u0438 \"\u0412\u044B\u0431\u0440\u0430\u0442\u044C \u0432\
            \u0441\u0435\" - \u0441\u043E\u0437\u0434\u0430\u0435\u0442 \u0432\u0441\
            \u0435 \u043A\u043E\u043B\u043B\u0430\u0436\u0438 \u0441\u0440\u0430\u0437\
            \u0443\n\nArgs:\n    batch_size: \u0420\u0430\u0437\u043C\u0435\u0440\
            \ \u043F\u0430\u043A\u0435\u0442\u0430 (\u043F\u043E \u0443\u043C\u043E\
            \u043B\u0447\u0430\u043D\u0438\u044E 12)\n\nReturns:\n    JSON \u0441\
            \ \u0438\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\u0435\u0439 \u043E\
            \ \u0432\u0441\u0435\u0445 \u0441\u043E\u0437\u0434\u0430\u043D\u043D\u044B\
            \u0445 \u043A\u043E\u043B\u043B\u0430\u0436\u0430\u0445"
         operationId: create_all_collage_info_collage_batch_all__get
         parameters:
         -  in: query
            name: batch_size
            required: false
            schema:
               default: 12
               maximum: 12
               minimum: 1
               title: Batch Size
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create All Collage Info
         tags:
         - collages
   /collage/batch/all/download/:
      get:
         description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u0432\u0441\u0435\
            \ \u043A\u043E\u043B\u043B\u0430\u0436\u0438 \u0438\u0437 \u0432\u0441\
            \u0435\u0445 \u0442\u043E\u0432\u0430\u0440\u043E\u0432 \u0438 \u043E\u0442\
            \u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442 \u0438\u0445 \u043F\u043E\
            \u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044E \u0432 \u0432\
            \u0438\u0434\u0435 ZIP-\u0430\u0440\u0445\u0438\u0432\u0430\n\u0418\u0441\
            \u043F\u043E\u043B\u044C\u0437\u0443\u0435\u0442\u0441\u044F \u0434\u043B\
            \u044F \u043A\u043D\u043E\u043F\u043A\u0438 \"\u0412\u044B\u0431\u0440\
            \u0430\u0442\u044C \u0432\u0441\u0435\" - \u0441\u043E\u0437\u0434\u0430\
            \u0435\u0442 \u0432\u0441\u0435 \u043A\u043E\u043B\u043B\u0430\u0436\u0438\
            \ \u0438 \u0441\u0440\u0430\u0437\u0443 \u043E\u0442\u043F\u0440\u0430\
            \u0432\u043B\u044F\u0435\u0442 \u0434\u043B\u044F \u0441\u043A\u0430\u0447\
            \u0438\u0432\u0430\u043D\u0438\u044F\n\nArgs:\n    batch_size: \u0420\u0430\
            \u0437\u043C\u0435\u0440 \u043F\u0430\u043A\u0435\u0442\u0430 (\u043F\u043E\
            \ \u0443\u043C\u043E\u043B\u0447\u0430\u043D\u0438\u044E 12)\n\nReturns:\n\
            \    ZIP-\u0430\u0440\u0445\u0438\u0432 \u0441\u043E \u0432\u0441\u0435\
            \u043C\u0438 \u0441\u043E\u0437\u0434\u0430\u043D\u043D\u044B\u043C\u0438\
            \ \u043A\u043E\u043B\u043B\u0430\u0436\u0430\u043C\u0438"
         operationId: create_and_dowload_all_collages_collage_batch_all_download__get
         parameters:
         -  in: query
            name: batch_size
            required: false
            schema:
               default: 12
               maximum: 12
               minimum: 1
               title: Batch Size
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Create And Dowload All Collages
         tags:
         - collages
   /collage/batch/info/:
      get:
         description: "\u0412\u043E\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442\
            \ \u0438\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\u044E \u043E\
            \ \u0442\u0435\u043A\u0443\u0449\u0435\u043C \u043F\u0430\u043A\u0435\u0442\
            \u0435 \u0442\u043E\u0432\u0430\u0440\u043E\u0432 \u0431\u0435\u0437 \u0441\
            \u043E\u0437\u0434\u0430\u043D\u0438\u044F \u043A\u043E\u043B\u043B\u0430\
            \u0436\u0430\n\u0418\u0441\u043F\u043E\u043B\u044C\u0437\u0443\u0435\u0442\
            \u0441\u044F \u0434\u043B\u044F \u043F\u043E\u043B\u0443\u0447\u0435\u043D\
            \u0438\u044F \u0438\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\u0438\
            \ \u043F\u0435\u0440\u0435\u0434 \u0441\u043E\u0437\u0434\u0430\u043D\u0438\
            \u0435\u043C \u043A\u043E\u043B\u043B\u0430\u0436\u0430\n\nArgs:\n   \
            \ start_index: \u041D\u0430\u0447\u0430\u043B\u044C\u043D\u044B\u0439\
            \ \u0438\u043D\u0434\u0435\u043A\u0441 \u0434\u043B\u044F \u043E\u0431\
            \u0440\u0430\u0431\u043E\u0442\u043A\u0438\n    batch_size: \u0420\u0430\
            \u0437\u043C\u0435\u0440 \u043F\u0430\u043A\u0435\u0442\u0430 (\u043F\u043E\
            \ \u0443\u043C\u043E\u043B\u0447\u0430\u043D\u0438\u044E 12)\n\nReturns:\n\
            \    JSON \u0441 \u0438\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\
            \u0435\u0439 \u043E \u0442\u0435\u043A\u0443\u0449\u0435\u043C \u043F\u0430\
            \u043A\u0435\u0442\u0435"
         operationId: get_batch_info_collage_batch_info__get
         parameters:
         -  in: query
            name: start_index
            required: false
            schema:
               default: 0
               minimum: 0
               title: Start Index
               type: integer
         -  in: query
            name: batch_size
            required: false
            schema:
               default: 12
               maximum: 12
               minimum: 1
               title: Batch Size
               type: integer
         responses:
            '200':
               content:
                  application/json:
                     schema: {}
               description: Successful Response
            '422':
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/HTTPValidationError'
               description: Validation Error
         summary: Get Batch Info
         tags:
         - collages
...